Extract components/widgets and organize packages in a clean architecture way
